Service class and controller

+++ src/main/java/com/assignment/banking/BankingService/model/MortgageRate.java
@@ -3,13 +3,12 @@ package com.assignment.banking.BankingService.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
 public class MortgageRate {
     private int maturityPeriod;
-    private BigDecimal interestRate;
+    private double interestRate;
     private LocalDateTime lastUpdate;
 }
